requirements.txt
coverage==3.7.1
Django==1.8.5
django-bootstrap-toolkit==2.15.0
django-crispy-forms==1.5.2
django-filter==0.11.0
django-guardian==1.3.2
django-registration-redux==1.1
djangorestframework==3.3.2
Markdown==2.6.5
Pillow==2.6.1
Pygments==2.0.2
six==1.10.0
wheel==0.24.0
